Fix dosdebug (ldt, symbol reading) and cpu exception printf for 64-bit.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1609 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/include/mhpdbg.h
@@ -145,7 +145,7 @@ extern int restart_cputime (int);
 #endif
 
 struct symbol_entry {
-   unsigned int  addr;
+   unsigned long addr;
    unsigned char type;
    unsigned char name[49];
 };
